Minor fixes and tweaks
